+++ antlr/test.sh
@@ -10,22 +10,30 @@
 # - uncomment one of the test commands below. Feel free to add new commands.
 # - run this script like this: . test.sh
 
+jarfile=antlr-4.9.2-complete.jar
+libdir=./lib
+jarpath=$libdir/$jarfile
+
+export CLASSPATH=".:./out:./$libdir/$jarfile"
+alias grun='java -Dfile.encoding=UTF-8 org.antlr.v4.gui.TestRig'
+
 #########################################################
 # Helpful commands for testing the StateSmith Grammar:
 #########################################################
-# grun StateSmithLabelGrammar node -gui -tokens -trace test_input.txt.txt
-# grun StateSmithLabelGrammar node -tokens -trace test_input.txt.txt | grep "consume"
-# grun StateSmithLabelGrammar node -gui -tokens -trace test_input.txt.txt
-# grun StateSmithLabelGrammar node -tokens -trace test_input.txt.txt > /dev/null
+# grun StateSmithLabelGrammar node -gui -tokens -trace test_input.txt
+grun StateSmithLabelGrammar node -tokens test_input.txt
+# grun StateSmithLabelGrammar node -tokens -trace test_input.txt | grep "consume"
+# grun StateSmithLabelGrammar node -gui -tokens -trace test_input.txt
+# grun StateSmithLabelGrammar node -tokens -trace test_input.txt > /dev/null
 
 #########################################################
 # Helpful commands for testing the PlantUML Grammar:
 #########################################################
 ## parses and shows Parse Tree Inspector GUI
-grun PlantUML diagram -gui -tokens -trace test_input.txt
+# grun PlantUML diagram -gui -tokens -trace test_input.txt
 
 ## parses and shows which tokens are consumed
-grun PlantUML diagram -tokens -trace test_input.txt | grep "consume"
+# grun PlantUML diagram -tokens -trace test_input.txt | grep "consume"
 
 ## if you are looking for an error, this will show only stderr
 # grun PlantUML diagram -tokens -trace test_input.txt > /dev/null