+++ src/StateSmith.Cli/CHANGELOG.md
@@ -8,6 +8,19 @@ Test/interim releases are not documented here.
 
 ---
 
+## [0.18.?]
+## Added
+- `run` - Allow specifying input file pattern(s) instead of relying on directory scanning or manifest.
+  - ex: `ss.cli run sm1.plantuml`, `ss.cli run **.plantuml **.drawio*`
+  - globbing is cross-platform
+  - https://github.com/StateSmith/StateSmith/issues/422
+
+### Changed (minor)
+- `ss.cli run --verbose` now prints all parsed CLI args.
+- `ss.cli run` - minor console message tweaks.
+
+---
+
 ## [0.17.5]
 ### Contributors
 - @emmby